==> 0_Reddit_project_3_testing.ipynb <==
import requests
import time

import pandas as pd

url_donald_g = 'https://www.reddit.com/r/thedonald.json'

url_Donald_T = 'https://www.reddit.com/r/The_Donald/.json'

headers = {'User-agent': 'James Baker'}


res_1 = requests.get(url_Donald_T,headers = headers)
res_1.status_code

res_2 = requests.get(url_donald_g,headers = headers)
res_2.status_code

content = res_2.json()

content

content.keys()

content['data'].keys()

content['data']['children'][1].keys()

[content['data']['children'][i]['data']['name'] for i in range (25)]

content['data']['children'][15]['data']['selftext']

[content['data']['children'][i]['data']['selftext'] for i in range (25)]

trump = res_1.json()

[trump['data']['children'][i]['data']['title'] for i in range (25)]

[trump['data']['children'][i]['data']['selftext'] for i in range (25)]

dg_posts = []
after = None
for i in range(40):
    print(i)
    if after == None:
        params = {}
    else:
        params = {'after':after}
    url = 'https://www.reddit.com/r/thedonald.json' #donald glover
    res = requests.get(url,params=params,headers=headers)
    if res.status_code == 200:
        the_json = res.json()
        dg_posts.extend(the_json['data']['children'])
        after = the_json['data']['after']
    else:
        print(res.status_code)
        break 
    time.sleep(1)
dg_data = pd.DataFrame(dg_posts)
dg_data.to_csv('glover.csv2')

djt_posts = []
after = None
for i in range(40):
    print(i)
    if after == None:
        params = {}
    else:
        params = {'after':after}
    url = 'https://www.reddit.com/r/The_Donald/.json' #Donald Trump
    res = requests.get(url,params=params,headers=headers)
    if res.status_code == 200:
        the_json = res.json()
        djt_posts.extend(the_json['data']['children'])
        after = the_json['data']['after']
    else:
        print(res.status_code)
        break 
    time.sleep(1)
djt_data = pd.DataFrame(djt_posts)
djt_data.to_csv('trump.csv2')
