==> 1_BC_scraper.ipynb <==
#using Brian Collins code to get all the data from the pushshift api.
#tried a few subreddits before settling on 'The_Donald' and 'Showerthoughts'

import pandas as pd
import numpy as np
import requests
import time
import datetime as dt
import dateutil

def query_pushshift(subreddit, kind='submission', skip=1, times=21, 
                    subfield = ['title', 'selftext', 'subreddit', 'created_utc', 'author', 'num_comments', 'score', 'is_self'],
                    comfields = ['body', 'score', 'created_utc']):

    stem = "https://api.pushshift.io/reddit/search/{}/?subreddit={}&size=500".format(kind, subreddit)
    mylist = []
    
    for x in range(1, times):
        
        URL = "{}&after={}d".format(stem, skip * x)
        print(URL)
        response = requests.get(URL)
        assert response.status_code == 200
        mine = response.json()['data']
        df = pd.DataFrame.from_dict(mine)
        mylist.append(df)
        time.sleep(2)
        
    full = pd.concat(mylist, sort=False)
    
    if kind == "submission":
        
        full = full[subfield]
        
        #full = full.drop_duplicates()
        
#         full = full.loc[full['is_self'] == True]
        
    def get_date(created):
        return dt.date.fromtimestamp(created)
    
    _timestamp = full["created_utc"].apply(get_date)
    
    full['timestamp'] = _timestamp

    print(full.shape)
    
    return full 

DJT = query_pushshift('The_Donald')

len(DJT['title'].value_counts().values)

#DJT.drop_duplicates(subset='title')

DJT.to_csv('/Users/zacharystern/Documents/GA/submissions/Project_3_reddit/trump_f.csv')

glover = query_pushshift('thedonald', times=50)

len(glover['title'].value_counts().values)

pd.options.display.max_rows = 999

#glover.sort_values('title')

glover.to_csv('/Users/zacharystern/Documents/GA/submissions/Project_3_reddit.glover_f.csv')

shower = query_pushshift('Showerthoughts', times=25)

len(shower['title'].value_counts().values)

shower.to_csv('/Users/zacharystern/Documents/GA/submissions/Project_3_reddit/shower.csv')

Other to check https://www.reddit.com/r/EnoughTrumpSpam/
