==> 2_Project_3_start.ipynb <==
#import, clean data, combine into one dataset

import pandas as pd
import numpy as np
import re


#change column width option
pd.set_option('display.max_colwidth',100)

#read in the 'shower thoughts' posts:
shower = pd.read_csv('./data/shower.csv')

shower.head()

#creating a list of columns that we will drop and not use in our model
drop = [col for col in shower.columns if col not in ['title','subreddit']]

shower.drop(drop,axis=1,inplace=True)

#checking the shape of our new, cleaned df
shower.shape

#dropping charicters that are not needed and will interfer w/ our model
shower['title'] = shower['title'].map(lambda x: x.lower())

shower['title'] = shower['title'].map(lambda x: re.sub('r/', '', x)  )

shower['title'] = shower['title'].map(lambda x: re.sub(r'[$\"&<.,%*?>😎]', '',x))

shower['title'] = shower['title'].map(lambda x: re.sub(r'\d*','',x))

shower.head()

#reading in our Trump aka 'r_donald' reddit postings
trump = pd.read_csv('./data/trump_f.csv')

trump.head()

#running the same cleaning operations below as we did for our first data frame

trump.drop(drop,axis=1,inplace=True)

trump['title'] = trump['title'].map(lambda x: x.lower())
trump['title'] = trump['title'].map(lambda x: re.sub('r/', '', x)  )
trump['title'] = trump['title'].map(lambda x: re.sub(r'[$\"&<.,%*?>😎]', '',x))
trump['title'] = trump['title'].map(lambda x: re.sub(r'\d*', '',x))

#check shape on trump posts
trump.shape

#deleting/dropping some 'shower thoughts' posts so that we have equal classes
shower.drop(shower.index[10000:12000],inplace=True)

shower.shape

#create combined final df
final = shower.append(trump)

final.shape

#checking the distribution of our classes prior to prediction
final['subreddit'].value_counts(normalize=True)

#write our dataframe to csv to save
#final.to_csv('./data/reddit_final.csv')
