==> 4_Project_3_model-NO_SHOWER.ipynb <==
#trying out removing of possible identifying keywords from posts such as "trump' or 'shower'

import pandas as pd
import numpy as np
import re

from sklearn.model_selection import train_test_split,cross_val_score,GridSearchCV,KFold

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.pipeline import Pipeline
from sklearn.naive_bayes import MultinomialNB


data = pd.read_csv('./data/reddit_final.csv')

data.drop('Unnamed: 0',axis=1,inplace=True)

data.isnull().sum()

data.dropna(axis=0,inplace=True)

data.head()

data['subreddit'].value_counts(normalize=True)

#Set shower thoughts as target 1, trump as 0
targets = {'Showerthoughts':1,'The_Donald':0}

data['target'] = data['subreddit'].map(targets)

data['target'].value_counts(normalize=True)

data[['target','subreddit']].head()

#checking on the keywords in each subreddit catagory

data[data['title'].str.contains("shower")].head()

data[(data['subreddit']=='The_Donald') & (data['title'].str.contains('trump'))].head()

#remove Trump,trump, and shower related terms

data['title'] = data['title'].map(
    lambda x: re.sub(r'([Ss]hower|[Ss]howerthoughts|[Tt]rump|[Dd]onald)','',x)
)

data[data['title'].str.contains('trump')]

X = data['title']
y = data['target']

X_train, X_test, y_train, y_test = train_test_split(X,y,random_state=37,stratify=y)

kf = KFold(n_splits=3,shuffle=True)

cv = CountVectorizer(stop_words='english',max_features=6000,ngram_range=(1,3))

train_cv = cv.fit_transform(X_train)

test_cv = cv.transform(X_test)

print(train_cv.shape)
print(test_cv.shape)

lrg = LogisticRegression(penalty='l2',C=1)

cross_val_score(lrg,train_cv.toarray(),y_train,cv=kf).mean()

lrg.fit(train_cv.toarray(),y_train)

lrg.score(train_cv.toarray(),y_train)

lrg.score(test_cv.toarray(),y_test)

df = pd.DataFrame(train_cv.toarray(),columns=cv.get_feature_names())

df.head()

train_cv.toarray().sum(axis=0);

rf = RandomForestClassifier()

rf.fit(train_cv.toarray(),y_train)

print(rf.score(train_cv.toarray(),y_train))
print(rf.score(test_cv.toarray(),y_test))

nb = MultinomialNB()

cross_val_score(nb,train_cv.toarray(),y_train,cv=kf).mean()

rf_params = {
    'n_estimators': [250,300,450],
    #'max_depth': [200,400,800],
    

    
    
}

# gs = GridSearchCV(rf,param_grid=rf_params,n_jobs=-2)

# gs.fit(train_cv.toarray(), y_train)
# print(gs.best_score_)
# gs.best_params_
